job-seekers-reach-out-to-hiring-managers/agentic/autogen/pyproject.toml
[project]
name = "job-seekers-autogen"
version = "1.0.0"
description = "Job seekers reach out to hiring managers using AutoGen and ConnectSafely.ai"
authors = [
    { name = "ConnectSafely.ai", email = "support@connectsafely.ai" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "autogen-agentchat>=0.7.0",
    "autogen-ext[google,openai]>=0.4.0",
    "google-genai>=0.2.0",
    "streamlit>=1.39.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
